Update sufia.en.yml
Terminology edited by Jake Carlson on Feb 11, 2016.

Some remaining issues: directory suffix (line 19 & 20) has example@org. I was not sure what to replace this with.
lines 50-54, I was not sure if we have particular sources (such as Box) to list here, or if leaving this generic "cloud providers" made more sense.

lines 134, 137 - It's not clear to me what the user is "updating". Is this for instances of the user saving a draft during the upload process and returning to it later? Once the submission is made the user would not be able to edit it, right? Also, I don't know that a failure to update would be due to the user lacking sufficient privileges to edit. I'm inclined to remove this, but wanted to ask you first.

line 140 - are we replacing the default background image with one of our own? If so, we need to change this to UM or whatever the source of the image would be. 
 
Lines 143-147 - I don't believe that we allowing for users to have "private" items and so I think this could be deleted.

Lines 149, 151 - I would remove these. No one likes to be told that they have no followers. 

Lines 179 - 182 - We ask asking for more metadata in the deposit form than what is being listed here. I'm assuming one field for method and one field for description would be fine. We will need to work on adding metadata for dates. There is the date of creation and the dates of coverage field; we should give the user the opportunity to add additional dates for each.

Line 182 - Remove? I believe that we are handling rights through asking the user to select a Creative Commons license from a list of options.

Line 195 - Is this needed? Can we drop it? 

Lines 208 - 223 - I'm not sure what the "simple form" refers to. Searching the collections in DBD? If so, we would need to modify the labels to match up with the metadata we are collecting from depositors.

Line 225 - 240 - We have decided on required metadata: title, creator (author), date created, method, description, and license. I don't know that we have decided on what (if any) optional metadata we would want to include or make available. All of the items listed on these lines could potentially be useful and so I did not edit them. I will return to this once we have a better sense of possibilities for optional metadata.